// sourcecode/BTrak.Packages/snova-payroll/module/module-components/src/lib/Payroll/models/PayRollTemplateConfigurationModel.ts
export class PayRollTemplateConfigurationModel{
    payRollTemplateConfigurationId: string;
    payRollComponentId: string;
    payRollTemplateId: string;
    isPercentage: boolean;
    amount: number;
    percentagevalue: number;
    isCtcDependent: boolean;
    isRelatedToPT: boolean;
    payRollComponentName: string;
    searchText: string;
    isArchived: boolean;
    timeStamp:any;
    componentName: string;
    isNew:boolean;
    payRollComponentIds: any[];
    type: number;
    dependentType: number;
    componentId: string;
    isDeduction:  boolean;
    dependentPayRollComponentId: string;
    dependentPayRollComponentName: string;
    order: number;
    relatedToContributionPercentage:  boolean;
    isBands:  boolean;
}