// sourcecode/BTrak.Packages/snova-payroll/module/module-components/src/lib/Payroll/models/PayRollComponentModel.ts
export class PayRollComponentModel{
    payRollComponentId: string;
    componentName: string
    isDeduction: boolean;
    isVariablePay: boolean;
    isArchived: boolean;
    timeStamp: any;
    employeeContributionPercentage: number;
    employerContributionPercentage: number;
    isVisible: boolean;
    relatedToContributionPercentage : boolean;
}